requirements.txt
Chameleon==2.24
Mako==1.0.3
MarkupSafe==0.23
PasteDeploy==1.5.2
Pygments==2.0.2
SQLAlchemy==1.0.9
WebOb==1.5.1
argparse==1.4.0
cornice==1.1.0
google-api-python-client==1.4.2
linecache2==1.0.0
nodeenv==0.13.6
ordereddict==1.1
pyramid==1.5.7
pyramid-chameleon==0.3
pyramid-debugtoolbar==2.4.2
pyramid-mako==1.0.2
pyramid-tm==0.12
repoze.lru==0.6
simplejson==3.8.1
six==1.10.0
traceback2==1.4.0
transaction==1.4.4
translationstring==1.3
unittest2==1.1.0
venusian==1.0
waitress==0.8.10
wsgiref==0.1.2
zope.deprecation==4.1.2
zope.interface==4.1.3
zope.sqlalchemy==0.7.6
